nursesTimetable/pyproject.toml
[tool.poetry]
name = "nursestimetable"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.6"
django = "^4.2"
requests = "^2.32.3"
djangorestframework = "^3.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"